requirements/base.txt
pyyaml==5.3.*
